+++ src/Web/Framework/Plzwrk/Util.hs
@@ -29,7 +29,7 @@ import           Web.Framework.Plzwrk.Base      ( dats
                                                 , dats'
                                                 , PwAttribute(..)
                                                 )
-import           Web.Framework.Plzwrk.Browserful
+import           Web.Framework.Plzwrk.JSEnv
 
 -- | Creates a text attribute wrapped in an applicative functor
 pT :: String -> s -> PwAttribute s opq
@@ -41,7 +41,7 @@ pF f _ = PwFunctionAttribute f
 
 -- |From an event, gets the target's value.
 eventTargetValue
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the event
   -> IO (Maybe String) -- ^ the target value, or nothing if it doesn't exist
 eventTargetValue browser e = do
@@ -50,7 +50,7 @@ eventTargetValue browser e = do
 
 -- |From an event, takes the target and blurs it.
 eventTargetBlur
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval  -- ^ the event
   -> IO () -- ^ returns nothing
 eventTargetBlur browser e = do
@@ -59,7 +59,7 @@ eventTargetBlur browser e = do
 
 -- |Take an event and prevent the default.
 eventPreventDefault
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the event
   -> IO () -- ^ returns nothing
 eventPreventDefault browser e = void $ invokeOn0 browser e "preventDefault"
@@ -68,7 +68,7 @@ eventPreventDefault browser e = void $ invokeOn0 browser e "preventDefault"
 
 -- | Sets on an element an attribute. See [Element.setAttribute](https://developer.mozilla.org/en-US/docs/Web/API/Element/setAttribute)
 elementSetAttribute
-  :: Browserful jsval -- ^ The browser
+  :: JSEnv jsval -- ^ The browser
   -> jsval -- ^ the node
   -> String -- ^ the attribute name
   -> String -- ^ the attribute
@@ -80,7 +80,7 @@ elementSetAttribute b e k v = do
 
 -- | Gets the tag name of an element.  See [Element.tagName](https://developer.mozilla.org/en-US/docs/Web/API/Element/tagName)
 elementTagName
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the element
   -> IO (Maybe String) -- ^ Returns the tag name
 elementTagName b v = do
@@ -89,7 +89,7 @@ elementTagName b v = do
 
 -- | Takes a target and an event name and adds a listener. See [EventTarget.addEventListener](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener)
 eventTargetAddEventListener
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the element
   -> String -- ^ the listener name. note that this should be "click" or "input", not "onclick" nor "oninput"
   -> jsval -- ^ the listener
@@ -100,7 +100,7 @@ eventTargetAddEventListener b e k v = do
 
 -- | Takes a target and an event name and removes a listener. See [EventTarget.removeEventListener](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/removeEventListener)
 eventTargetRemoveEventListener
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the element
   -> String -- ^ the listener name. note that this should be "click" or "input", not "onclick" nor "oninput"
   -> jsval -- ^ the listener
@@ -111,7 +111,7 @@ eventTargetRemoveEventListener b e k v = do
 
 -- | Gets a JavaScript property as a bool, returning @Nothing@ if the object being called is null or undefined or the property cannot be cast to a bool.
 getPropertyAsBool
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the object containing the property
   -> String -- ^ the property name
   -> IO (Maybe Bool) -- ^ the response if the property is a bool, else Nothing
@@ -121,7 +121,7 @@ getPropertyAsBool b o k = do
 
 -- | Gets a JavaScript property as a double, returning @Nothing@ if the object being called is null or undefined or the property cannot be cast to a double.
 getPropertyAsDouble
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the object containing the property
   -> String -- ^ the property name
   -> IO (Maybe Double) -- ^ the response if the property is a double, else Nothing
@@ -131,7 +131,7 @@ getPropertyAsDouble b o k = do
 
 -- | Gets a JavaScript property as an int, returning @Nothing@ if the object being called is null or undefined or the property cannot be cast to an int.
 getPropertyAsInt
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the object containing the property
   -> String -- ^ the property name
   -> IO (Maybe Int) -- ^ the response if the property is an int, else Nothing
@@ -141,7 +141,7 @@ getPropertyAsInt b o k = do
 
 -- | Gets a JavaScript property as an string, returning @Nothing@ if the object being called is null or undefined.
 getPropertyAsString
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the object containing the property
   -> String -- ^ the property name
   -> IO (Maybe String) -- ^ the response
@@ -150,18 +150,18 @@ getPropertyAsString b o k = do
   maybe (pure Nothing) (castToString b) _v
 
 -- | Takes an element and clicks it. Useful for testing. See [HTMLElement.click](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/click)
-htmlElemenetClick :: Browserful jsval -> jsval -> IO ()
+htmlElemenetClick :: JSEnv jsval -> jsval -> IO ()
 htmlElemenetClick b e = void $ invokeOn0 b e "click"
 
 -- | Logs a string. See [Console.log](https://developer.mozilla.org/en-US/docs/Web/API/Console/log)
-consoleLogS :: Browserful jsval -> String -> IO ()
+consoleLogS :: JSEnv jsval -> String -> IO ()
 consoleLogS b s = do
   _s <- jsValFromString b s
   consoleLog b _s
 
 -- | Takes a node and appends a child. See [Node.appendChild](https://developer.mozilla.org/en-US/docs/Web/API/Node/appendChild)
 nodeAppendChild
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the node
   -> jsval -- ^ the child to append
   -> IO () -- ^ returns nothing
@@ -169,7 +169,7 @@ nodeAppendChild b e v = void $ invokeOn1 b e "appendChild" v
 
 -- | Get the children of a node. See [Node.childNodes](https://developer.mozilla.org/en-US/docs/Web/API/Node/childNodes)
 nodeChildNodes
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the node
   -> IO (Maybe [jsval])
 nodeChildNodes b v = do
@@ -178,7 +178,7 @@ nodeChildNodes b v = do
 
 -- | Inserts a node into an element before another node.  See [Node.insertBefore](https://developer.mozilla.org/en-US/docs/Web/API/Node/insertBefore)
 nodeInsertBefore
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the parent element
   -> jsval -- ^ the new node
   -> jsval -- ^ the pre-existing node
@@ -187,7 +187,7 @@ nodeInsertBefore b e k v = void $ invokeOn2 b e "insertBefore" k v
 
 -- | Removes a child from a parent node.  See [Node.removeChild](https://developer.mozilla.org/en-US/docs/Web/API/Node/removeChild)
 nodeRemoveChild
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the parent element
   -> jsval -- ^ the child to remove
   -> IO () -- ^ returns nothing
@@ -195,7 +195,7 @@ nodeRemoveChild b e v = void $ invokeOn1 b e "removeChild" v
 
 -- | Gets the text content of a node. See [Node.textContent](https://developer.mozilla.org/en-US/docs/Web/API/Node/textContent)
 nodeTextContent
-  :: Browserful jsval -- ^ the browser
+  :: JSEnv jsval -- ^ the browser
   -> jsval -- ^ the node
   -> IO (Maybe String) -- ^ the text content as a string
 nodeTextContent b e = do